requirements.txt
selenium==3.141.0
urllib3==1.26.6
PyYAML==5.4.1
